+++ draw.py
@@ -1,7 +1,10 @@
 import bpy
 import blf
 import gpu
+
+from mathutils import Vector
 from gpu_extras.batch import batch_for_shader
+# from .preferences import get_addon_prefs
 
 ## Draw utils
 
@@ -39,7 +42,85 @@ def stop_callback(self, context):
     context.window.cursor_set("DEFAULT")
     if hasattr(self, '_handle'):
         bpy.types.SpaceView3D.draw_handler_remove(self._handle, 'WINDOW')
+    if hasattr(self, '_pos_handle'):
+        bpy.types.SpaceView3D.draw_handler_remove(self._pos_handle, 'WINDOW')
     context.area.tag_redraw()
+
+
+def origin_position_callback(self, context):
+    # Draw origin position VS Ghost old position
+    # self.objects
+    # self.init_mats[i]
+    
+    ## TODO: Handle case with one objects and case with multiple objects
+
+    coords = [] # Origin coords
+    ghost_coords = [] # Starting origin coords
+
+    cross_coord = [
+        Vector((-0.1, 0, 0)), Vector((0.1, 0, 0)),
+        Vector((0, -0.1, 0)), Vector((0, 0.1, 0)),
+        Vector((0, 0, -0.1)), Vector((0, 0, 0.1)),
+    ]
+
+    if hasattr(self, 'objects'):
+        ## Multi object
+        for ob in self.objects:
+            coords += [ob.matrix_world @ v for v in cross_coord]
+        
+        # if hasattr(self, 'init_mats')
+    elif hasattr(self, 'object'):
+        ## Single object
+        coords = [self.object.matrix_world @ v for v in cross_coord]
+    elif hasattr(self, 'ob'):
+        ## Single object
+        coords = [self.ob.matrix_world @ v for v in cross_coord]
+    else:
+        return
+
+    gpu.state.blend_set('ALPHA')
+    gpu.state.line_width_set(2)
+
+    if bpy.app.version <= (3,6,0):
+        shader = gpu.shader.from_builtin('3D_UNIFORM_COLOR')
+    else:
+        shader = gpu.shader.from_builtin('UNIFORM_COLOR')
+    
+
+    origin_batch = batch_for_shader(shader, 'LINES', {"pos": coords})
+
+    if hasattr(self, 'init_mats'):
+        ## Multi object
+        for mat in self.init_mats:
+            ghost_coords += [mat @ v for v in cross_coord]
+    elif hasattr(self, 'init_mat'):
+        ## Single object
+        ghost_coords = [self.init_mat @ v for v in cross_coord]
+
+    if ghost_coords:
+        ## Draw ghost
+        ghost_batch = batch_for_shader(shader, 'LINES', {"pos": ghost_coords})
+        ghost_color = (0.5, 0.5, 0.5, 0.3)
+        shader.uniform_float("color", ghost_color)
+        ghost_batch.draw(shader)
+    
+    
+    ## Draw current origin AFTER ghost to display on top
+
+    ## Dimmed version always visible (no depth test)
+    shader.uniform_float("color", (1.0, 0.6, 0.3, 0.2))
+    origin_batch.draw(shader)
+
+    ## Redraw with more vivid color WITH depth test
+    previous_depth_test_value = gpu.state.depth_test_get()
+    gpu.state.depth_test_set('LESS')
+    shader.uniform_float("color", (1.0, 0.6, 0.3, 0.8)) # orange bright
+    origin_batch.draw(shader)
+    
+    gpu.state.depth_test_set(previous_depth_test_value)
+
+    gpu.state.line_width_set(1)
+    gpu.state.blend_set('NONE')
 
 
 ## Not used yet